// src/Web/reactprojectTreeEtalon/src/Library/Abstract3DConverters/MeshCreators/Obj3DCreator.d.ts
import type { IFactory } from "../../Interfaces/IFactory";
import type { IMtlDetector } from "../Interfaces/IMtlDetector";
import type { ITextureIndex } from "../Interfaces/ITextureIndex";
import type { IMesh } from "../Interfaces/IMesh";
import type { ITextReaderFactory } from "../../IO/Interfaces/ITextReaderFactory";
import { ColorTexture } from "../ColorTexture";
import { EffectTexture } from "../EffectTexture";
import { ImageTexture } from "../ImageTexture";
import { LinesMeshCreator } from "./LinesMeshCreator";
export declare class Obj3DCreator extends LinesMeshCreator {
    constructor(url: string, name: string, directory: string, obj: any, factory: IFactory, func: ITextReaderFactory | undefined);
    getIndexes(): ITextureIndex[][];
    private pushIndex;
    getNames(): string[];
    loadLines(): void;
    private iindexes;
    private static siindexes;
    getVertices(): number[][];
    getTextures(): number[][];
    getNormals(): number[][];
    getDefaultEffect(): EffectTexture;
    getMeshCreatorMeshes(): IMesh[];
    createMeshes(): void;
    createMaterialsFromLUrl(url: string, eff: EffectTexture[]): Map<string, EffectTexture>;
    getMeshName(): string;
    efff: EffectTexture[];
    createMaterialsFromLines(lines: string[], eff: EffectTexture[]): Map<string, EffectTexture>;
    getMeshCreatorEffects(): Map<string, EffectTexture>;
    createEffect(f: string): EffectTexture;
    createGeometry(): void;
    mttt: Map<string, EffectTexture>;
    createMaterials(): void;
    createEffectFromImage(f: string): EffectTexture;
    detect(st: string): EffectTexture | undefined;
    getInitial(line: string): string | undefined;
    createDefaultGeometry(): void;
    createNamedGeometry(): void;
    createUnNamedGeometry(): void;
    effectsPrivate: Map<string, EffectTexture>;
    ka: ImageTexture;
    kd: ImageTexture;
    ks: ImageTexture;
    ambient: ColorTexture;
    emissive: ColorTexture;
    specular: ColorTexture;
    ns: number;
    ni: number;
    d: number;
    illum: number;
    effect: EffectTexture;
    default: EffectTexture;
    nm: number;
    usedMaterials: string[];
    objs: string;
    fiction: string;
    names: string[];
    tuple: ITextureIndex;
    mtlDetetctor: IMtlDetector;
    materialLines: string[];
    mtll: string;
}
//# sourceMappingURL=Obj3DCreator.d.ts.map